+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to `Laravel Package Skeleton` will be documented in this fil
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## V0.3 Alpha  07-05-2019
+
+- Fix artisan
+- Add Skeleton Make Command
+- Add composer file creation and run
+- Add provider file creation and add register configuration
+
+## V0.2 Alpha  06-05-2019
+
+- fix composer packages are not dev
+
+
 ## V0.1 Alpha  06-05-2019
 
 - Initial testing release